Move OkHttpSocketConnection to okhttp module
