+++ jellyfin-api-okhttp/api/jellyfin-api-okhttp.api
@@ -20,3 +20,11 @@ public final class org/jellyfin/sdk/api/okhttp/OkHttpFactory : org/jellyfin/sdk/
 	public final fun getBase ()Lokhttp3/OkHttpClient;
 }
 
+public final class org/jellyfin/sdk/api/okhttp/OkHttpSocketConnection : org/jellyfin/sdk/api/sockets/SocketConnection {
+	public fun <init> (Lokhttp3/OkHttpClient;Lkotlinx/coroutines/CoroutineScope;)V
+	public fun connect (Ljava/lang/String;Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun disconnect (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun getState ()Lkotlinx/coroutines/flow/StateFlow;
+	public fun send (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
