+++ jellyfin-api/api/jellyfin-api.api
@@ -1264,14 +1264,6 @@ public final class org/jellyfin/sdk/api/sockets/DefaultSocketApi : org/jellyfin/
 	public fun subscribeAll ()Lkotlinx/coroutines/flow/Flow;
 }
 
-public final class org/jellyfin/sdk/api/sockets/OkHttpSocketConnection : org/jellyfin/sdk/api/sockets/SocketConnection {
-	public fun <init> (Lokhttp3/OkHttpClient;Lkotlinx/coroutines/CoroutineScope;)V
-	public fun connect (Ljava/lang/String;Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun disconnect (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun getState ()Lkotlinx/coroutines/flow/StateFlow;
-	public fun send (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
 public abstract interface class org/jellyfin/sdk/api/sockets/SocketApi {
 	public abstract fun getState ()Lkotlinx/coroutines/flow/StateFlow;
 	public abstract fun subscribe (Lkotlin/reflect/KClass;)Lkotlinx/coroutines/flow/Flow;
